refactor: add features folder and cart feature
