+++ src/features/cart/DTOs/cart.types.ts
@@ -1,4 +1,5 @@
 import type { AdminCoupon } from "store/adminCoupon/adminCoupon.types";
+
 export interface CartItems {
   id: string;
   coupon: AdminCoupon;
