+++ src/features/cart/components/cart-icon/CartIcon.styles.scss
@@ -1,4 +1,4 @@
-@use "../../stylesheets/animation";
+@use "../../../../stylesheets/animation";
 
 .cart-icon {
   display: flex;
@@ -27,4 +27,3 @@
     right: -20%;
   }
 }
-
