+++ src/features/cart/hooks/cart.hooks.ts
@@ -1,0 +1,24 @@
+import { useState, useEffect } from "react";
+import { useAppSelector } from "../../../store/redux-hooks";
+import { selectCartItemsQuantity } from "../store/cart.selector";
+
+export const useCartDetectedQuantity = () => {
+  const [isItemChange, setIsItemChange] = useState(false);
+  const quantity = useAppSelector(selectCartItemsQuantity);
+
+  useEffect(() => {
+    if (quantity === 0) {
+      return;
+    }
+    setIsItemChange(true);
+    const timer = setTimeout(() => {
+      setIsItemChange(false);
+    }, 300);
+
+    return () => {
+      clearTimeout(timer);
+    };
+  }, [quantity]);
+
+  return { quantity, isItemChange, setIsItemChange };
+};
