+++ src/features/cart/index.ts
@@ -1,0 +1,6 @@
+export * from "./components/cart-icon/CartIcon.component";
+export * from "./DTOs/cart.types";
+export * from "./store/cart.asyncThunk";
+export * from "./store/cart.selector";
+export * from "./store/cart.slice";
+export * from "./hooks/cart.hooks";
