+++ src/features/cart/store/cart.selector.ts
@@ -1,6 +1,6 @@
 import { createSelector } from "@reduxjs/toolkit";
 
-import type { RootState } from "../store";
+import type { RootState } from "../../../store/store";
 
 const selectCartReducer = (state: RootState) => state.cart;
